move agents stroage init to agent to get he possibilities to create agent config there
