+++ package.json
@@ -76,6 +76,7 @@
     "express": "^4.21.2",
     "langchain": "^0.3.34",
     "langsmith": "^0.3.73",
+    "openevals": "^0.1.1",
     "pg": "^8.15.6",
     "prom-client": "^15.1.3",
     "socket.io": "^4.8.1",
