+++ packages/agent/src/agents/graphs/core-graph/supervisor.graph.ts
@@ -1,8 +1,5 @@
-import { DynamicStructuredTool } from '@langchain/core/tools';
 import { CompiledStateGraph, StateGraph } from '@langchain/langgraph';
 import { PostgresSaver } from '@langchain/langgraph-checkpoint-postgres';
-import { Postgres } from '@snakagent/database';
-import { AnyZodObject } from 'zod';
 import { GraphError } from '../utils/error.utils.js';
 import { SupervisorAgent } from '@agents/core/supervisorAgent.js';
 import { skipValidationType } from '@stypes/graph.types.js';
@@ -204,7 +201,6 @@ export class SupervisorGraph {
       preModelHook: this.transformMessagesHook.bind(this),
       postModelHook: this.addAditionalKwargsToMessage.bind(this),
     });
-
     return workflow;
   }
 }
